dcpy reorg
